Add AboutPage component
