+++ src/components/organisms/PrivateRoute.jsx
@@ -1,16 +1,20 @@
 import React from "react";
 import { useSelector } from "react-redux";
-import { Redirect, Route } from "react-router-dom";
+import { Redirect, Route, useLocation } from "react-router-dom";
 
 const PrivateRoute = ({ component: Component, ...rest }) => {
   const { token } = useSelector(state => state.currentUser);
+  const { pathname } = useLocation();
+
   return (
     <Route
       {...rest}
       render={({ location, ...props }) =>
-        !token 
+        token 
         ? <Component {...props} /> 
-        : <Redirect to={{ pathname: "/register", state: { from: location.pathname } }} />
+        : pathname !== '/register' 
+        ? <Redirect to={{ pathname: "/register", state: { from: location.pathname } }} />
+        : null
       }
     />
   );
